moved dropseq_cmd and picard_cmd to be methods of Config, which feels a little weird but makes some sense
