+++ src/slideseq/config.py
@@ -2,6 +2,7 @@ import importlib.resources
 import logging
 from dataclasses import dataclass
 from pathlib import Path
+from typing import Union
 
 import yaml
 
@@ -36,6 +37,66 @@ class Config:
             gsheet_id=data["gsheet_id"],
             worksheet=data["worksheet"],
         )
+
+    def dropseq_cmd(
+        self,
+        command: str,
+        input_file: Union[Path, str],
+        output_file: Union[Path, str],
+        tmp_dir: Path,
+        mem: str = "8g",
+        compression: int = 0,
+    ):
+        """Return the beginning of a DropSeq command, with standard options
+
+        :param command: name of the dropseq tool being invoked
+        :param input_file: path to the input file
+        :param output_file: path to the output file
+        :param tmp_dir: Location of the tmp directory to use
+        :param mem: memory for the heap. default is to share with other jobs
+        :param compression: compression level for output. Use 0 for speed, 5 for storage
+        """
+
+        return [
+            self.dropseq_dir / command,
+            "-m",
+            mem,
+            f"I={input_file}",
+            f"O={output_file}",
+            f"TMP_DIR={tmp_dir}",
+            "VALIDATION_STRINGENCY=SILENT",
+            f"COMPRESSION_LEVEL={compression}",
+            "VERBOSITY=WARNING",
+            "QUIET=true",
+        ]
+
+    def picard_cmd(self, command: str, tmp_dir: Path, mem: str = "62g"):
+        """Return the beginning of a Picard command, with standard options
+
+        :param command: name of the picard tool being invoked
+        :param tmp_dir: Location of the tmp directory to use
+        :param mem: Memory for the heap. Lower this for piped commands
+        """
+        return [
+            "java",
+            f"-Djava.io.tmp_dir={tmp_dir}",
+            f"-Xms{mem}",
+            f"-Xmx{mem}",
+            "-XX:+UseParallelGC",
+            "-XX:GCTimeLimit=20",
+            "-XX:GCHeapFreeLimit=10",
+            "-jar",
+            self.picard,
+            command,
+            "--TMP_DIR",
+            tmp_dir,
+            "--VALIDATION_STRINGENCY",
+            "SILENT",
+            "--VERBOSITY",
+            "WARNING",
+            "--QUIET",
+            "true",
+        ]
 
 
 def get_config() -> Config:
