+++ src/slideseq/pipeline/alignment.py
@@ -9,13 +9,7 @@ import click
 from slideseq.alignment_quality import write_alignment_stats
 from slideseq.config import get_config
 from slideseq.metadata import Manifest
-from slideseq.util import (
-    dropseq_cmd,
-    give_group_access,
-    picard_cmd,
-    run_command,
-    start_popen,
-)
+from slideseq.util import give_group_access, run_command, start_popen
 from slideseq.util.logger import create_logger
 
 log = logging.getLogger(__name__)
@@ -68,8 +62,7 @@ def main(
 
     procs = []
 
-    cmd = dropseq_cmd(
-        config.dropseq_dir,
+    cmd = config.dropseq_cmd(
         "TagBamWithReadSequenceExtended",
         library.raw_ubam,
         "/dev/stdout",
@@ -97,12 +90,8 @@ def main(
     )
 
     # Tag bam with read sequence extended molecular
-    cmd = dropseq_cmd(
-        config.dropseq_dir,
-        "TagBamWithReadSequenceExtended",
-        "/dev/stdin",
-        "/dev/stdout",
-        manifest.tmp_dir,
+    cmd = config.dropseq_cmd(
+        "TagBamWithReadSequenceExtended", "/dev/stdin", "/dev/stdout", manifest.tmp_dir
     )
     cmd.extend(
         [
@@ -126,20 +115,14 @@ def main(
     )
 
     # Filter low-quality reads
-    cmd = dropseq_cmd(
-        config.dropseq_dir, "FilterBam", "/dev/stdin", "/dev/stdout", manifest.tmp_dir
-    )
+    cmd = config.dropseq_cmd("FilterBam", "/dev/stdin", "/dev/stdout", manifest.tmp_dir)
     cmd.extend(["PASSING_READ_THRESHOLD=0.1", "TAG_REJECT=XQ"])
 
     procs.append(start_popen(cmd, "FilterBam", library, lane, procs[-1]))
 
     # Trim reads with starting sequence
-    cmd = dropseq_cmd(
-        config.dropseq_dir,
-        "TrimStartingSequence",
-        "/dev/stdin",
-        "/dev/stdout",
-        manifest.tmp_dir,
+    cmd = config.dropseq_cmd(
+        "TrimStartingSequence", "/dev/stdin", "/dev/stdout", manifest.tmp_dir
     )
     cmd.extend(
         [
@@ -153,12 +136,8 @@ def main(
     procs.append(start_popen(cmd, "TrimStartingSequence", library, lane, procs[-1]))
 
     # Adapter-aware poly A trimming
-    cmd = dropseq_cmd(
-        config.dropseq_dir,
-        "PolyATrimmer",
-        "/dev/stdin",
-        library.polya_filtered_ubam,
-        manifest.tmp_dir,
+    cmd = config.dropseq_cmd(
+        "PolyATrimmer", "/dev/stdin", library.polya_filtered_ubam, manifest.tmp_dir
     )
     cmd.extend(
         [
@@ -180,7 +159,7 @@ def main(
     log.debug("Finished with pre-alignment processing")
 
     # convert to fastq like a loser
-    cmd = picard_cmd(config.picard, "SamToFastq", manifest.tmp_dir)
+    cmd = config.picard_cmd("SamToFastq", manifest.tmp_dir)
     cmd.extend(
         [
             "-I",
@@ -227,7 +206,7 @@ def main(
     procs = []
 
     # Sort aligned bam
-    cmd = picard_cmd(config.picard, "SortSam", manifest.tmp_dir, mem="24g")
+    cmd = config.picard_cmd("SortSam", manifest.tmp_dir, mem="24g")
     cmd.extend(
         [
             "-I",
@@ -241,7 +220,7 @@ def main(
     procs.append(start_popen(cmd, "SortSam", library, lane))
 
     # Merge unmapped bam and aligned bam
-    cmd = picard_cmd(config.picard, "MergeBamAlignment", manifest.tmp_dir, mem="24g")
+    cmd = config.picard_cmd("MergeBamAlignment", manifest.tmp_dir, mem="24g")
     cmd.extend(
         [
             "-R",
@@ -264,20 +243,15 @@ def main(
     procs.append(start_popen(cmd, "MergeBamAlignment", library, lane, procs[-1]))
 
     # Tag read with interval
-    cmd = dropseq_cmd(
-        config.dropseq_dir,
-        "TagReadWithInterval",
-        "/dev/stdin",
-        "/dev/stdout",
-        manifest.tmp_dir,
+    cmd = config.dropseq_cmd(
+        "TagReadWithInterval", "/dev/stdin", "/dev/stdout", manifest.tmp_dir
     )
     cmd.extend([f"INTERVALS={library.reference.intervals}", "TAG=XG"])
 
     procs.append(start_popen(cmd, "TagReadWithInterval", library, lane, procs[-1]))
 
     # Tag read with gene function
-    cmd = dropseq_cmd(
-        config.dropseq_dir,
+    cmd = config.dropseq_cmd(
         "TagReadWithGeneFunction",
         "/dev/stdin",
         library.processed_bam,
